+++ cron_expression.go
@@ -7,24 +7,25 @@ package cron
 import (
 	"context"
 
+	"github.com/bborbe/run"
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
 	"github.com/robfig/cron/v3"
 )
 
 func NewExpressionCron(
-	ext string,
-	action action,
+	expression string,
+	action run.Runnable,
 ) CronJob {
 	return &cronExpression{
-		ext:    ext,
-		action: action,
+		expression: expression,
+		action:     action,
 	}
 }
 
 type cronExpression struct {
-	ext    string
-	action action
+	expression string
+	action     run.Runnable
 }
 
 func (c *cronExpression) Run(ctx context.Context) error {
@@ -32,19 +33,18 @@ func (c *cronExpression) Run(ctx context.Context) error {
 	errChan := make(chan error)
 	cronJob := cron.New()
 	parser := cron.NewParser(cron.Second | cron.Minute | cron.Hour | cron.Dom | cron.Month | cron.Dow | cron.Descriptor)
-	schedule, err := parser.Parse(c.ext)
+	schedule, err := parser.Parse(c.expression)
 	if err != nil {
-		return errors.Wrapf(err, "parse cron expression '%s' failed", c.ext)
+		return errors.Wrapf(err, "parse cron expression '%s' failed", c.expression)
 	}
 
 	cronJob.Start()
 	job := cron.FuncJob(func() {
 		glog.V(4).Infof("run cron action started")
-		if err := c.action(ctx); err != nil {
-			glog.V(2).Infof("action failed -> exit")
+		if err := c.action.Run(ctx); err != nil {
 			errChan <- err
 		}
-		glog.V(4).Infof("run cron action finished")
+		glog.V(4).Infof("run cron action completed")
 	})
 	id := cronJob.Schedule(schedule, job)
 	glog.V(3).Infof("scheduled job: %v", id)