+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
+## v1.1.0
+
+- refactor
+- return context cancel error
+- use run.Runnable as action
+
 ## v1.0.1
 
 - refactor